+++ core/util.lua
@@ -1,8 +1,5 @@
 local ADDON_NAME, ns = ...
 
-local L = ns.locale
-local Class = ns.Class
-
 -------------------------------------------------------------------------------
 ------------------------------ DATAMINE TOOLTIP -------------------------------
 -------------------------------------------------------------------------------
@@ -270,64 +267,6 @@ local function HEXtoRGBA(color)
 end
 
 -------------------------------------------------------------------------------
-------------------------------- Interval Class --------------------------------
--------------------------------------------------------------------------------
-
-local Interval = Class('Interval')
-
-function Interval:Initialize(attrs)
-    if attrs then for k, v in pairs(attrs) do self[k] = v end end
-
-    local region_initial = {
-        [1] = self.initial.us,
-        [2] = self.initial.kr or self.initial.tw,
-        [3] = self.initial.eu,
-        [5] = self.initial.cn
-    } -- https://wowpedia.fandom.com/wiki/API_GetCurrentRegion
-
-    if self.id then
-        self.SpawnTime = self.id * self.offset +
-                             (region_initial[GetCurrentRegion()] or
-                                 self.initial.us)
-    end
-end
-
-function Interval:Next()
-    if not (self.id and self.initial and self.interval) then return false end
-    local CurrentTime = GetServerTime()
-    local SpawnTime = self.SpawnTime
-
-    local NextSpawn = SpawnTime +
-                          math.ceil((CurrentTime - SpawnTime) / self.interval) *
-                          self.interval
-    local TimeLeft = NextSpawn - CurrentTime
-
-    return NextSpawn, TimeLeft
-end
-
-function Interval:GetText()
-    local TimeFormat = ns:GetOpt('use_standard_time') and self.format_12hrs or
-                           self.format_24hrs
-
-    local NextSpawn, TimeLeft = self:Next()
-
-    local SpawnsIn = TimeLeft <= 60 and L['now'] or
-                         SecondsToTime(TimeLeft, true, true)
-
-    if self.yellow and self.green then
-        local color = ns.color.Orange
-        if TimeLeft < self.yellow then color = ns.color.Yellow end
-        if TimeLeft < self.green then color = ns.color.Green end
-        SpawnsIn = color(SpawnsIn)
-    end
-
-    local text = format('%s (%s)', SpawnsIn, date(TimeFormat, NextSpawn))
-    if self.text then text = format(self.text, text) end
-    ns.PrepareLinks(text)
-    return text
-end
-
--------------------------------------------------------------------------------
 
 ns.AsIDTable = AsIDTable
 ns.AsTable = AsTable
@@ -339,4 +278,3 @@ ns.PlayerHasItem = PlayerHasItem
 ns.PlayerHasProfession = PlayerHasProfession
 ns.PrepareLinks = PrepareLinks
 ns.RenderLinks = RenderLinks
-ns.Interval = Interval