+++ plugins/10_Dragonflight/localization/koKR.lua
@@ -220,6 +220,7 @@ L['options_icons_tuskarr_chest'] = nil
 L['options_icons_tuskarr_chest_desc'] = nil
 
 L['community_feast_label'] = nil
+L['tasks_completed_suffix'] = nil
 L['options_icons_community_feast'] = nil
 L['options_icons_community_feast_desc'] = nil
 
@@ -234,6 +235,13 @@ L['options_icons_icemaw_storage_cache_desc'] = nil
 L['kazzi_note_start'] = nil
 L['kazzi_note_item'] = nil
 L['kazzi_achievement_suffix'] = nil
+
+L['naszuro_vakthros'] = nil
+L['naszuro_imbu'] = nil
+L['naszuro_azure_archives'] = nil
+L['naszuro_hudsons_rock'] = nil
+
+L['ferry_to_iskaara'] = nil
 
 -------------------------------------------------------------------------------
 ------------------ FORBIDDEN REACH (DRACTHYR STARTING ZONE) -------------------
@@ -357,6 +365,8 @@ L['renown_envoy_label'] = nil
 L['renown_envoy_note'] = nil
 L['trader_hagarth_note'] = nil
 
+L['naszuro_caldera_of_the_menders'] = nil
+
 -------------------------------------------------------------------------------
 ------------------------------ OHN'AHRAN PLAINS -------------------------------
 -------------------------------------------------------------------------------
@@ -460,6 +470,9 @@ L['options_icons_lightning_bound_chest_desc'] = nil
 
 L['bloodgullet_note'] = nil
 
+L['naszuro_windsong_rise'] = nil
+L['naszuro_emerald_gardens'] = nil
+
 -------------------------------------------------------------------------------
 --------------------------------- THALDRASZUS ---------------------------------
 -------------------------------------------------------------------------------
@@ -529,6 +542,11 @@ L['options_icons_titan_chest_desc'] = nil
 L['living_mud_mask_note'] = nil
 
 L['little_scales_daycare_note'] = nil
+L['naszuro_veiled_ossuary'] = nil
+L['naszuro_algethar_academy'] = nil
+L['naszuro_vault_of_the_incarnates'] = nil
+L['naszuro_thaldraszus_peak'] = nil
+L['naszuro_temporal_conflux'] = nil
 
 -------------------------------------------------------------------------------
 ------------------------------ THE WAKING SHORE -------------------------------
@@ -668,6 +686,11 @@ L['phoenix_wishwing_info'] = nil
 
 L['bugbiter_tortoise_note'] = nil
 
+L['naszuro_apex_canopy'] = nil
+L['naszuro_obsidian_throne'] = nil
+L['naszuro_ruby_lifeshrine'] = nil
+L['naszuro_dragonheart_outpost'] = nil
+
 -------------------------------------------------------------------------------
 ------------------------------- Zaralek Cavern --------------------------------
 -------------------------------------------------------------------------------
@@ -702,6 +725,7 @@ L['molten_scoutbot_note'] = nil
 L['bolts_and_brass_note'] = nil
 
 L['sniffen_sage_suffix'] = nil
+L['sniffen_digs_suffix'] = nil
 
 L['big_slick_note'] = nil
 L['grogul_note'] = nil